Fix routes by standard rest
