+++ src/routes/index.js
@@ -1,10 +1,9 @@
 const { Router } = require('express');
-const restaurantRoute = require('./restaurantRoute');
-const neighborhoodRoute = require('./neighborhoodRoute');
-const userRoute = require('./userRoute');
+const restaurantsRoute = require('./restaurantsRoute');
+const neighborhoodsRoute = require('./neighborhoodsRoute');
+const usersRoute = require('./usersRoute');
 const cuisinesRoute = require('./cuisinesRoute');
 const loginRoute = require('./loginRoute');
-const reviewsRoute = require('./reviewsRoute');
 const reservesRoute = require('./reservesRoute');
 // Importar todos los routers;
 // Ejemplo: const authRouter = require('./auth.js');
@@ -12,12 +11,11 @@ const router = Router();
 
 // Configurar los routers
 // // Ejemplo: router.use('/auth', authRouter);
-router.use('/restaurant', restaurantRoute)
-router.use('/user', userRoute)
-router.use('/neighborhood', neighborhoodRoute)
+router.use('/restaurants', restaurantsRoute)
+router.use('/users', usersRoute)
+router.use('/neighborhoods', neighborhoodsRoute)
 router.use('/cuisines', cuisinesRoute)
-router.use('/login', loginRoute)
-router.use('/review', reviewsRoute)
-router.use('/reserve', reservesRoute)
+router.use('/logins', loginRoute)
+router.use('/reserves', reservesRoute)
 
 module.exports = router;
