+++ src/routes/usersRoute.js
@@ -1,9 +1,10 @@
 const express = require("express");
-const { User, Restaurant } = require("../db");
+const { User, Restaurant, Review, Reserve } = require("../db");
 const bcrypt = require("bcryptjs");
 
 const router = express.Router();
 
+//Creo un usuario
 router.post("/", async (req, res) => {
   const { username, email, password, password2 } = req.body;
 
@@ -48,6 +49,7 @@ router.post("/", async (req, res) => {
   }
 });
 
+//Obtengo los restaurants creados por un usuario en particular
 router.get("/:id/restaurants", async (req, res) => {
   const { id } = req.params;
 
@@ -78,6 +80,66 @@ router.get("/:id/restaurants", async (req, res) => {
     }
   } catch (e) {
     return res.status(404).json({ message: "Petición inválida" });
+  }
+});
+
+//Obtengo las reviews creadas por un usuario en particular
+router.get("/:id/reviews", async (req, res) => {
+  //id de usuario
+  const { id } = req.params;
+
+  try {
+    if (id) {
+      const reviews = await Review.findAll({
+        where: {
+          UserId: id,
+        },
+      });
+      if (reviews.length) {
+        // console.log(reviews);
+        return res.status(200).send(reviews);
+      } else {
+        return res
+          .status(200)
+          .json({ message: "El usuario no ha realizado ninguna reseña" });
+      }
+    } else {
+      return res.status(400).json({
+        message: "Hace falta el ID del usuario para encontrar sus reseñas",
+      });
+    }
+  } catch (e) {
+    return res.status(400).json({ message: "Petición inválida" })
+  }
+});
+
+//Obtengo todas las reservas realizadas por un usuario en particular
+router.get("/:id/reserves", async (req, res) => {
+  //id de usuario
+  const { id } = req.params;
+
+  try {
+    if (id) {
+      const reserves = await Reserve.findAll({
+        where: {
+          UserId: id,
+        },
+      });
+      if (reserves.length) {
+        // console.log(reserves);
+        return res.status(200).send(reserves);
+      } else {
+        return res.status(200).json({
+          message: "El usuario no tiene reservas para los próximos días",
+        });
+      }
+    } else {
+      return res.status(400).json({
+        message: "Hace falta el ID del usuario para encontrar sus reservas",
+      });
+    }
+  } catch (e) {
+    console.log(e);
   }
 });
